refactor(frontend): トップページの初回画面をWelcomeSectionとして再構成し、Chevronアイコンを共通化
- タイピング演出コンポーネントを WelcomeSection から Greeting にリネーム
  （実装手段ではなく中身の役割で命名し、外側の合成レイヤーが WelcomeSection を名乗れるようにする）
- top/page.tsx に直書きだった初回画面の合成（Greeting + プロフィール導線 + スクロール誘導 +
  ビューポート高レイアウト）を新設の WelcomeSection に切り出し、他ページと同じ
  「名前のついたセクションを並べるだけ」の構成に揃える
- プロフィール導線コンポーネントを ProfileSection から ProfileTeaser にリネーム
  （入れ子になった時点でページ直下のセクション単位ではなくなったことと、
  /ui/profile ページ側の ProfileDetailSection 等との混同を避けるため）
- 下向き/上向きの山形アイコンを ChevronIcon として切り出し、
  CareerItem の開閉ヒントと WelcomeSection のスクロール誘導で共通利用する
  （上向きは rotate-180 の付与だけで表現し、2つの path を持たせない）
