+++ frontend/src/app/components/ChevronIcon.tsx
@@ -1,0 +1,20 @@
+type ChevronIconProps = {
+  className?: string;
+};
+
+export default function ChevronIcon({ className }: ChevronIconProps) {
+  return (
+    <svg
+      className={className}
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+    >
+      <path d="M6 9l6 6 6-6" />
+    </svg>
+  );
+}
